Feat(#200) Move BuildingParticularRisk validator to its own file.
